requirements.txt
json5 == 0.8.5
numpy == 1.16.4
torch == 1.1.0
tqdm == 4.32.1
transformers == 2.5.1
sklearn == 0.0
PyYAML
